Add enable/disable guessing using client IP/port
Add configurable option for whether to include client port or client IP
in the protocol guesses. This defaults to include both client port/IP
if the protocol is not guessed with the server IP/port.

This is intended for when the flow direction detection is enabled, so we
know that sport = client port, dport = server port.